+++ packages/interactions/types/index.d.ts
@@ -1,0 +1,5 @@
+import active from "./active";
+import ancestors from "./ancestors";
+import pathname from "./pathname";
+import prefetch from "./prefetch";
+export { active, ancestors, pathname, prefetch };