+++ packages/router/CHANGELOG.md
@@ -3,9 +3,8 @@
 * `router.route` returns a route's public data.
 * Restructure a route's public data to have `meta`, `methods`, and `extra` properties.
 * Rework interactions to be external, ideally stateless, functions.
-* Merge in the `ancestors` interaction, which now returns an array of ancestor routes.
-* Merge in the `prefetch` interaction.
 * Removed `partials` property from response objects.
+* Re-export interactions from `@curi/interactions`.
 
 ## 2.0.0-beta.15